Attach on Event Tag Fix
When using tags, objects were found only in the Start function. This doesn't account for tagged objects being spawned in after the game is started, making this component essentially useless when using prefabs. The code was moved from Start to OnEvent, so it reassesses the object pool every time it receives the trigger event.